requirements.txt
pillow
numpy
scikit-learn
